Database gets reset in unit tests and fixing any remaining problems. Hence, the project is almost finished. Need to optimize view as there are still some problems with the offered keyboards in telegram. Closes Eng-MFQ/BBQReserverOfTeam3#9

+++ BBQReserverBot/BBQReserverBot/Model/RecordModel.cs
@@ -46,6 +46,7 @@ namespace BBQReserverBot.Model
 
         public static bool DeleteRecord(Record record)
         {
+            if (!DatabaseController.RecordExists(record)) return false;
             var command = "delete from records where id = " + record.Id;
             DatabaseController.ExecuteCommand(command);
             return true;
